+++ lib/core/modifier.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/widgets.dart';
+
+abstract class Modifier {
+  const Modifier();
+
+  Widget apply(Widget widget);
+}