+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* First create plugin
+
+## 0.0.2
+
+* Added button component and multiple modifiers